% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{build_filename}
\alias{build_filename}
\title{Build path to file}
\usage{
build_filename(
  filename = "example",
  extension = ".png",
  output = ".",
  datetime = F
)
}
\arguments{
\item{filename}{string. Name of the file, default="example".}

\item{extension}{string. File extension, default=".png".}

\item{output}{string. Output path, default=".".}

\item{datetime}{logical. Whether to append the datetime in format.
Default FALSE.}
}
\value{
String of the filename.
}
\description{
This function builds a complete path to file including:
output path + filename + extension.
}
\examples{
build_filename(filename = "example", extension = ".png", output = ".",
datetime = F)
build_filename(filename = "example", extension = ".png", output = ".",
datetime = T)
}
\author{
Luca Lalli, Stefano Bergamini
}
