% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContVarTest.R
\name{cont_var_test_LB}
\alias{cont_var_test_LB}
\title{Test for continuous variables splitted by categories}
\usage{
cont_var_test_LB(
  data,
  variables,
  paired = FALSE,
  group,
  dumb = FALSE,
  statistic = "{mean} ({sd})",
  ID = "ID",
  num_dec = 2,
  p.adjust.method = NULL,
  excel = F,
  excel_path = paste0(path_out, "/Results.xlsx"),
  telegram = "none"
)
}
\arguments{
\item{data}{dataframe}

\item{variables}{vector containing all variables of interest}

\item{paired}{FALSE/TRUE}

\item{group}{factor variable splitting the data}

\item{dumb}{FALSE are you dumb? Hope not}

\item{statistic}{Specifies summary statistics to display for each variable. Default = "(mean) (sd)".}

\item{ID}{ID variabl (Default = "ID")}

\item{num_dec}{Decimal number for mean and SD (Default = 2)}

\item{p.adjust.method}{correction method, a character string. Can be abbreviated.}

\item{excel}{export fuction results as multiple Excel sheets}

\item{excel_path}{path where you want your Excel}

\item{telegram}{send a telegram message}
}
\value{
Una lista con dataset
}
\description{
The most powerful function ever created. You can perform the 4 major tests and the posthoc tests for Friedman and Kruskal-Wallis.
If you are dumb (option dumb = T) you can also perform posthoc tests without correcting for test multiplicity.
Please do not try this at home/work and consider asking a statistician before performing any test.
Stored functions for statistic option are (median), (mean), (sd), (min), (max), (q1), (q3), (n), and (range).
}
\examples{
cont_var_test_LB(data = iris, variables = c("Sepal.Length", "Sepal.Width"), group = "Species", paired = FALSE)
}
