% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{Descrittive}
\alias{Descrittive}
\title{Function to build, starting from a dataset, the descriptive statistics of every variable}
\usage{
Descrittive(dataset, path = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{path}{where do you want it to be saved}
}
\description{
Function to build, starting from a dataset, the descriptive statistics of every variable
}
\author{
Luca Lalli
}
