% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{Distribution}
\alias{Distribution}
\title{Plot variable distribution}
\usage{
Distribution(data, var, split = FALSE, cutoff = NULL)
}
\arguments{
\item{data}{A dataframe containing the variable to plot.}

\item{var}{string. Name of the variable to plot.}

\item{split}{logical. If TRUE, the variable is split by a cut-off value.
Default=FALSE.}

\item{cutoff}{numeric. Cut-off value to split the variable. Default=NULL.}
}
\value{
A ggplot object.
}
\description{
This function outputs a ggplot showing the distribution of a single variable,
where points are optionally divided by a cut-off value (cutoff).
}
\examples{
Distribution(data = mtcars, var = "mpg", split = TRUE, cutoff = 23)
}
\author{
Luca Lalli, Stefano Bergamini
}
