% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_function_for_LandS.R
\name{formatz_p}
\alias{formatz_p}
\title{Function to get a formatted p-value for a number o a vector of numbers}
\usage{
formatz_p(value, digits = 4)
}
\arguments{
\item{value}{a number or a vector of numbers to be formatted}

\item{digits}{the number of decimal places (Default = 4)}
}
\value{
a number or a vector of numbers formatted with 4 digits
}
\description{
Function to get a formatted p-value for a number o a vector of numbers
}
\examples{
formatz_p(c(1.000, 0.75643242, 0.000032431, 0.00214))

}
