% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{Kmax_aim}
\alias{Kmax_aim}
\title{Function to print the histogram of the AIM::cv.cox.main output}
\usage{
Kmax_aim(kmax.cycle = kmax.cycle)
}
\arguments{
\item{kmax.cycle}{The vector of values of the best biomarkers}
}
\value{
an histogram
}
\description{
Function to print the histogram of the AIM::cv.cox.main output
}
\author{
Luca Lalli, Stefano Bergamini
}
