% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_descrittive_v2.R
\name{LL_Descrittive}
\alias{LL_Descrittive}
\title{Function to build, starting from a dataset, the descriptive statistics of every variable}
\usage{
LL_Descrittive(dataset, path = NULL)
}
\arguments{
\item{dataset}{dataframe}

\item{path}{where do you want it to be saved}
}
\description{
Function to build, starting from a dataset, the descriptive statistics of every variable
}
