% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL_Fisher_gt_flextable.R
\name{LL_fisher_gt_flex}
\alias{LL_fisher_gt_flex}
\title{Function to build coloumn marginal statistics and Fisher test}
\usage{
LL_fisher_gt_flex(data, row_var, col_var, label_row_var, label_col_var)
}
\arguments{
\item{data}{dataframe}

\item{row_var}{row variable}

\item{col_var}{column variable}

\item{label_row_var}{label for row}

\item{label_col_var}{label for column}
}
\description{
Function to build coloumn marginal statistics and Fisher test
}
