% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multivariate.R
\name{multivariate}
\alias{multivariate}
\title{Function to create a multivariate cph model with a vector of variables}
\usage{
multivariate(db, vars, ptime, pevent, dec_HR = 4)
}
\arguments{
\item{db}{A dataframe}

\item{vars}{Vector of variables to be included in the multivariate model}

\item{ptime}{Survival Time variable}

\item{pevent}{Event variable}

\item{dec_HR}{digits of HR (Default = 4)}
}
\value{
the multivariate model
}
\description{
Function to create a multivariate cph model with a vector of variables
}
\author{
Luca Lalli
}
