% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_Report.R
\name{Outlier_Report_LB}
\alias{Outlier_Report_LB}
\title{Function for outliers report in excel}
\usage{
Outlier_Report_LB(
  data,
  variables,
  ID = "ID",
  Group = "Time",
  k = 1.5,
  excel = FALSE,
  excel_path = "Output/Outlier Report.xlsx"
)
}
\arguments{
\item{data}{a dataframe}

\item{variables}{vector of variables to check}

\item{ID}{ID variable. Default = "ID"}

\item{Group}{Grouping variable. Default = "Time"}

\item{k}{Parameter between Q1-Q3 and IQR. Default = 1.5}

\item{excel}{Whether to create an excel file of the report}

\item{excel_path}{Path for the excel file}
}
\value{
a list of dataframe one of the global variable, the other for the variable stratified for the grouping variable
}
\description{
Function for outliers report in excel
}
