% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_function_for_LandS.R
\name{output.aim.f}
\alias{output.aim.f}
\title{Function to print the output of the AIM function with Biomarker, Direction and Cutoff as a data frame model}
\usage{
output.aim.f(res.index, aim.data)
}
\arguments{
\item{res.index}{An output from the AIM package function}

\item{aim.data}{Data where the function was run on}
}
\value{
A dataframe-like object
}
\description{
Function to print the output of the AIM function with Biomarker, Direction and Cutoff as a data frame model
}
