% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{PDF_print_LB}
\alias{PDF_print_LB}
\title{Function to print the PDF with the grid.arrange function}
\usage{
PDF_print_LB(plot_list, path_print = path_print, nrow = 8, ncol = 6)
}
\arguments{
\item{plot_list}{The list you want to be plotted}

\item{path_print}{The path where you want your PDF to be printed}

\item{nrow}{Rows of your grid}

\item{ncol}{Columns of your grid}
}
\value{
A pdf in the path_output
}
\description{
Function to print the PDF with the grid.arrange function
}
