% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_function_for_LandS.R
\name{posthoc_df_LB}
\alias{posthoc_df_LB}
\title{Function for Boxplot_LB}
\usage{
posthoc_df_LB(Test_results, data, group, threshold_posthoc, i)
}
\arguments{
\item{Test_results}{Test_results}

\item{data}{data}

\item{group}{group}

\item{threshold_posthoc}{threshold posthoc tests}

\item{i}{i}
}
\value{
nothing
}
\description{
Function for Boxplot_LB
}
