% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{Posthoc_lineplots}
\alias{Posthoc_lineplots}
\title{Title}
\usage{
Posthoc_lineplots(Test_results, data, time, threshold_posthoc, i)
}
\arguments{
\item{Test_results}{Test_results}

\item{data}{data}

\item{threshold_posthoc}{Threshold ph tests}

\item{i}{i}

\item{Time}{Time}
}
\description{
Title
}
\author{
Luca Lalli, Stefano Bergamini
}
