% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Progress_bar_LB.R
\name{Progress_bar_LB}
\alias{Progress_bar_LB}
\title{Nice progress bar for cycles with different visual outputs}
\usage{
Progress_bar_LB(
  current,
  total,
  start_time,
  bar_length = 90,
  bar_fill = "=",
  bar_void = " ",
  logfile = NULL
)
}
\arguments{
\item{current}{The current iteractive value in the cycle}

\item{total}{The total number of variables in the cycle}

\item{start_time}{Start time of the cycle}

\item{bar_length}{Bar width in the console}

\item{bar_fill}{The bar fill character}

\item{bar_void}{The empty bar character}

\item{logfile}{If not null e.g. 'Log.txt' prints the loop log}
}
\value{
Only visual output in the console
}
\description{
Nice progress bar for cycles with different visual outputs
}
