% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_gsea_map.R
\docType{data}
\name{prot2gene}
\alias{prot2gene}
\title{Mapping of Olink protein names to gene names}
\format{
A character vector.
}
\usage{
prot2gene
}
\description{
This mapping covers the following Olink 96 Panels:
Immuno-oncology, Inflammation.
}
\examples{
print(prot2gene)

}
\author{
Stefano Bergamini
}
\keyword{datasets}
