% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity_kinetics_LB.R
\name{Similarity_Kinetic_LB}
\alias{Similarity_Kinetic_LB}
\title{Function to compare similarity between two kinetics}
\usage{
Similarity_Kinetic_LB(
  df_long,
  variables,
  grouping_var = NULL,
  time_var = "Time",
  trend = "similar",
  n_perm = 10000,
  seed = 123,
  excel = "similarity_output.xlsx"
)
}
\arguments{
\item{df_long}{Dataframe in full long format}

\item{variables}{Vector of variables to compare}

\item{grouping_var}{If NULL consider the global trend and compare the variables, if a two-level variable is indicated compare all the stratified variables for this}

\item{time_var}{Time variable name in the dataset}

\item{trend}{If “similar” evaluates the similar kinetics of the curves, if “correlated” evaluates the correlation of the kinetics}

\item{n_perm}{Number of permutations}

\item{seed}{Seed to replicate permutations}

\item{excel}{Path to export excel}
}
\value{
Results and file .xlsx
}
\description{
Function to compare similarity between two kinetics
}
