% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/String_LL.R
\name{Stringa_LL}
\alias{Stringa_LL}
\title{Funzione che riceve in input le posizioni dei nomi di un dataframe e crea una stringa di tali nomi separati da virgola o da altro segno/simbolo}
\usage{
Stringa_LL(data, vet, sep = ",")
}
\arguments{
\item{data}{datafrane}

\item{vet}{vector of positions for names in the dataset}

\item{sep}{symbol to separate names from each other}
}
\description{
Funzione che riceve in input le posizioni dei nomi di un dataframe e crea una stringa di tali nomi separati da virgola o da altro segno/simbolo
}
