% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarise_numeric_vars_by_group}
\alias{summarise_numeric_vars_by_group}
\title{Summarise multiple numeric variables by group}
\usage{
summarise_numeric_vars_by_group(data, numeric_vars, var_cat, digits = 2)
}
\arguments{
\item{data}{A data frame containing the variables to summarise.}

\item{numeric_vars}{Character vector giving the names of numeric variables
to summarise.}

\item{var_cat}{Character string giving the name of the categorical grouping
variable.}

\item{digits}{Integer. Number of decimal places used to round the summary
statistics. Default is 2.}
}
\value{
A named list of data frames. Each element corresponds to one variable in
numeric_vars; element names are the variable names.
}
\description{
Applies `summary_by_group()` to multiple numeric variables and
returns a named list of summary tables.
}
\examples{
data_example <- data.frame(
  group = rep(c("A", "B"), each = 5),
  value1 = c(1, 2, 3, NA, 5, 10, 12, 13, 15, NA),
  value2 = c(5, 4, 3, 2, 1, 20, 22, NA, 25, 26)
)

summarise_numeric_vars_by_group(
  data = data_example,
  numeric_vars = c("value1", "value2"),
  var_cat = "group"
)

}
\author{
Luca Lalli
}
