% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{Sys_Time}
\alias{Sys_Time}
\title{Function to get the Sys.time() in a cute and nice format}
\usage{
Sys_Time()
}
\value{
The Sys.time() in a cute format
}
\description{
Function to get the Sys.time() in a cute and nice format
}
\author{
Luca Lalli, Stefano Bergamini
}
