% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_function_for_LandS.R
\name{Sys_Time_LB}
\alias{Sys_Time_LB}
\title{Function to get the Sys.time() in a cute and nice format}
\usage{
Sys_Time_LB()
}
\value{
The Sys.time() in a cute format
}
\description{
Function to get the Sys.time() in a cute and nice format
}
