% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapply.R
\name{Tapply_f}
\alias{Tapply_f}
\title{Function for an easy application of the tapply}
\usage{
Tapply_f(data, var_quant, var_cat, digits = 2)
}
\arguments{
\item{data}{dataframe}

\item{var_quant}{quantitative variable}

\item{var_cat}{categorial variable}

\item{digits}{digits to display}
}
\description{
Function for an easy application of the tapply
}
\author{
Luca Lalli
}
