% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_function_for_LandS.R
\name{telegram_mess_LB}
\alias{telegram_mess_LB}
\title{Function to send a Telegram message with BiostatUO9 bot. NB: must create a start_time before running it}
\usage{
telegram_mess_LB(
  dest = "both",
  script = 0,
  rm_start_time = TRUE,
  timestamp = TRUE
)
}
\arguments{
\item{dest}{Who is going to receive the message}

\item{script}{The title of the message}

\item{rm_start_time}{If you want the start_time item to be removed after the message is sent}

\item{timestamp}{Do you want the time in your message}
}
\value{
Nothing
}
\description{
Function to send a Telegram message with BiostatUO9 bot. NB: must create a start_time before running it
}
