% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_LL_v02.R
\name{univariate_LL}
\alias{univariate_LL}
\title{This function allows you to create the univariate regression model for a vector of variables}
\usage{
univariate_LL(db, vars, ptime, pevent, dec_HR = 4)
}
\arguments{
\item{db}{dataframe}

\item{vars}{vector with variables name}

\item{ptime}{Survival Time variable}

\item{pevent}{Event variable}

\item{dec_HR}{digits of HR (Default = 4)}
}
\value{
a dataframe with all univariate models
}
\description{
This function allows you to create the univariate regression model for a vector of variables
}
