% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_LB_02_08_2024.R
\name{Boxplot_LB}
\alias{Boxplot_LB}
\title{This function creates a list of boxplot}
\usage{
Boxplot_LB(
  data,
  variables,
  group,
  ID_lines = FALSE,
  Posthoc = FALSE,
  Point = F,
  Median_line = F,
  rm.outliers = F,
  alpha_box = 0.1,
  width_box = 0.2,
  size_median_line = 0.8,
  lwd_box = 0.1,
  lwd_ID_line = 0.2,
  alpha_ID_line = 0.3,
  alpha_point = 0.3,
  size_point = 0.3,
  Test_results = NULL,
  threshold_posthoc = 0.1,
  axis_y_title = NULL,
  axis_x_title = NULL,
  size_axis_x = 6,
  size_axis_y = 6,
  ID = "ID",
  legend_cod = NULL,
  breaks_axis_x = levels(data[, group]),
  labels_axis_x = levels(data[, group]),
  grid = TRUE,
  PPTX = FALSE,
  pptx_width = 7.5,
  pptx_height = 5.5,
  extra = F,
  extra_text = NULL,
  palette_boxplot = c("salmon", "royalblue", "forestgreen", "gold"),
  palette_title = "black",
  size_title = 12,
  target = paste0(path_out, "/Boxplot.pptx"),
  ratio = 1,
  telegram = "none"
)
}
\arguments{
\item{data}{dataframe}

\item{variables}{vector containing all variables of interest}

\item{group}{factor variable splitting the data}

\item{ID_lines}{whether to print the lines for paired observations}

\item{Posthoc}{whether to display Posthoc tests}

\item{Point}{whether to display observation points}

\item{Median_line}{whether to display the line connecting medians}

\item{rm.outliers}{whether to remove outliers from display}

\item{alpha_box}{alpha parameter for the boxes}

\item{width_box}{box's width}

\item{size_median_line}{median linewidth}

\item{lwd_box}{box linewidth}

\item{lwd_ID_line}{linewidth for paired observations}

\item{alpha_ID_line}{alpha for paired observations}

\item{alpha_point}{alpha for points}

\item{size_point}{size for points}

\item{Test_results}{dataframe for global and posthoc tests, see cont_var_test_LB}

\item{threshold_posthoc}{threshold for displaying posthoc tests}

\item{size_axis_x}{axis y title dimension}

\item{size_axis_y}{axis x title dimension}

\item{ID}{ID variable}

\item{telegram}{}
}
\value{
Una lista di boxplot
}
\description{
This function creates a list of boxplot
}
