% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{Distribution}
\alias{Distribution}
\title{Distribution function}
\usage{
Distribution(data, var, split = FALSE, split_rule = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{var}{variable}

\item{split}{does the variable need to be splitted}

\item{split_rule}{the splitting rule}
}
\value{
a plot with the variable
}
\description{
Distribution function
}
\examples{
Distribution(data = mtcars, var = "mpg", split = TRUE, split_rule = 23)
}
\author{
Luca Lalli, Stefano Bergamini
}
