% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{filename_LB}
\alias{filename_LB}
\title{This function returns the filename to be outputted}
\usage{
filename_LB(
  filename = "Prova",
  extention = ".png",
  output = path_output,
  datetime = F
)
}
\arguments{
\item{filename}{name of the file}

\item{extention}{file extention}

\item{output}{the main output path}

\item{datetime}{whether to print the datetime in a cute format}
}
\description{
This function returns the filename to be outputted
}
