% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM Survival Curves.R
\name{KM_LB}
\alias{KM_LB}
\title{This function allows to create a KM survival curve overall or splitted by a categorical variable}
\usage{
KM_LB(
  Event = "OS_EVENT",
  tEvent = "OS",
  strata = 1,
  data = data,
  title = "Prova",
  xlab = "Time in months",
  ylab = "Probaility of Surv",
  xlim = c(0, max(data[, tEvent], na.rm = T)),
  breaks_by = 3
)
}
\arguments{
\item{Event}{Event variable}

\item{tEvent}{Survival Time Variable}

\item{strata}{Variable to stratify (Default = 1)}

\item{data}{dataframe}

\item{title}{Graph title (Default = "Prova")}

\item{xlab}{x-axis title (Default = "Time in months")}

\item{ylab}{y-axis title (Default = "Probaility of Surv")}

\item{xlim}{limits of x-axis (Default is from 0 to maximum observed time)}

\item{breaks_by}{breaks of risk table(Default = 3)}
}
\value{
a KM graph
}
\description{
This function allows to create a KM survival curve overall or splitted by a categorical variable
}
