% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{Kmax_aim_LB}
\alias{Kmax_aim_LB}
\title{Function to print the histogram of the AIM::cv.cox.main output}
\usage{
Kmax_aim_LB(kmax.cycle = kmax.cycle)
}
\arguments{
\item{kmax.cycle}{The vector of values of the best biomarkers}
}
\value{
an histogram
}
\description{
Function to print the histogram of the AIM::cv.cox.main output
}
