% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinePlots_7.2024.R
\name{Lineplots_LB}
\alias{Lineplots_LB}
\title{Function to build the lineplots}
\usage{
Lineplots_LB(
  data,
  variables,
  time,
  group = 1,
  split = F,
  lw_reg = 1,
  size_point = 0.6,
  size_title = 12,
  col_title = 1,
  size_axis_x = 5,
  size_axis_y = 6,
  size_title_grid = 7,
  breaks = unique(data[, time]),
  label = unique(data[, time]),
  ylim = c(0.2, 0.8),
  Posthoc = F,
  Friedman = F,
  Test_results = Test_results,
  grid = T,
  ratio = 1,
  extra = F,
  extra_text = NULL,
  PPTX = F,
  pptx_width = 8.5,
  pptx_height = 5.5,
  threshold_posthoc = 0.1,
  check = F,
  target = paste0(path, "/file.pptx"),
  col_lines = c("salmon", "royalblue")
)
}
\arguments{
\item{data}{dataset}

\item{variables}{vector of variables}

\item{time}{x-axis variable}

\item{group}{factor variable to group}

\item{split}{whether to split in two windows the lines}

\item{lw_reg}{lwd of regression line}

\item{size_point}{size of points}

\item{size_title}{size of title}

\item{col_title}{colour of title}

\item{size_axis_x}{x-axis text size}

\item{size_axis_y}{y-axis text size}

\item{size_title_grid}{size of title in the grid}

\item{breaks}{breaks of x-axis}

\item{label}{labels of x-axis}

\item{ylim}{ylim to display in the graph}

\item{Posthoc}{whether to display posthoc tests}

\item{Friedman}{friedman overall test dataset}

\item{Test_results}{posthoc test dataset}

\item{grid}{whether to build a grid or a pptx}

\item{ratio}{graph ratio}

\item{extra}{do you want to add extra option?}

\item{extra_text}{write your additional options}

\item{PPTX}{whether to build a pptx or a grid}

\item{pptx_width}{inch}

\item{pptx_height}{inch}

\item{threshold_posthoc}{threshold to display posthoc brackets}

\item{check}{check the correctness of your graph}

\item{target}{where do you want your pptx to be saved}

\item{col_lines}{splitted lines colour}
}
\description{
Function to build the lineplots
}
