% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{New_Project_LB}
\alias{New_Project_LB}
\title{Function to create a new project in the default folder}
\usage{
New_Project_LB(name_project, pc = c("Luca", "Stefano"))
}
\arguments{
\item{name_project}{The name of the Project}

\item{pc}{Which pc are we operating}
}
\value{
Returns a folder in Projects with Analisi, Dati and Output subfolders
}
\description{
Function to create a new project in the default folder
}
