% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsurv_format.R
\name{Npsurv_format}
\alias{Npsurv_format}
\title{Function to get a cute format of npsurv output.}
\usage{
Npsurv_format(fit.npsurv)
}
\arguments{
\item{fit.npsurv}{A npsurv(Surv(time, event) ~ cov_factor, data) object}
}
\value{
A cute format of npsurv output
}
\description{
Function to get a cute format of npsurv output.
}
\author{
Luca Lalli
}
