% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olink_gsea_map.R
\name{olink_gsea_map}
\alias{olink_gsea_map}
\title{Function which prepares names for olink_pathway_enrichment}
\usage{
olink_gsea_map(data, test_results)
}
\arguments{
\item{data}{NPX data frame in long format with at least protein name (Assay), OlinkID, UniProt,SampleID, QC_Warning, NPX, and LOD}

\item{test_results}{a dataframe of statistical test results including Adjusted_pval and estimate columns.}
}
\value{
the two data frames with Assay column adjusted for the pathway analysis
}
\description{
This function changes the Assay column in order to map all the genes in the pathway enrichment analysis
}
