% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom_Function_for_LandS.R
\name{posthoc_df}
\alias{posthoc_df}
\title{Function for Boxplot}
\usage{
posthoc_df(Test_results, data, group, threshold_posthoc, i)
}
\arguments{
\item{Test_results}{Test_results}

\item{data}{data}

\item{group}{group}

\item{threshold_posthoc}{threshold posthoc tests}

\item{i}{i}
}
\value{
nothing
}
\description{
Function for Boxplot
}
\author{
Luca Lalli, Stefano Bergamini
}
