% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{Posthoc_lineplots_LB}
\alias{Posthoc_lineplots_LB}
\title{Title}
\usage{
Posthoc_lineplots_LB(Test_results, data, time, threshold_posthoc, i)
}
\arguments{
\item{Test_results}{Test_results}

\item{data}{data}

\item{threshold_posthoc}{Threshold ph tests}

\item{i}{i}

\item{Time}{Time}
}
\description{
Title
}
