% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{Print_LB}
\alias{Print_LB}
\title{Function to print the PDF with the grid.arrange function}
\usage{
Print_LB(
  plot_list,
  path_print = path_print,
  nrow = 8,
  ncol = 6,
  ext = NULL,
  width_pg = 21,
  height_pg = 29.7,
  return = FALSE
)
}
\arguments{
\item{plot_list}{The list you want to be plotted}

\item{path_print}{The path where you want your PDF to be printed}

\item{nrow}{Rows of your grid}

\item{ncol}{Columns of your grid}

\item{ext}{File extention}

\item{width_pg}{page width in cm}

\item{height_pg}{page height in cm}

\item{return}{if you want to assign your grid}

\item{warning}{if you want to suppress warnings (Default = TRUE)}
}
\value{
A pdf in the path_output
}
\description{
Function to print the PDF with the grid.arrange function
}
