% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{Pushover_LB}
\alias{Pushover_LB}
\title{Sending a Pushover notification on your device}
\usage{
Pushover_LB(
  dest = "both",
  script = 0,
  timestamp = TRUE,
  priority = 0,
  app = "ayje1n4x8fi64bupdn5shjnwd8ut95",
  title = "RStudio",
  attachment = NULL,
  start_time = NULL
)
}
\arguments{
\item{dest}{Who is going to receive the notification (Default = both)}

\item{script}{Text of your message}

\item{timestamp}{If you want the time printed in your notification, if TRUE it requires a start_time in the .GlobalEnv}

\item{priority}{Priority of your notification (-2; 2)}

\item{app}{Your app API key}

\item{title}{Your title notification}

\item{attachment}{Path to an image, if you want to attach it}

\item{start_time}{Start time (Default = NULL)}
}
\description{
Sending a Pushover notification on your device
}
