% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom fucntion for LandS.R
\name{vett.quoted}
\alias{vett.quoted}
\title{Funzione che permette partendo da un vettore, di riscrivere quel vettore in varie forme}
\usage{
vett.quoted(vettore, sym = ", ", quote = T)
}
\arguments{
\item{vettore}{Starting vector}

\item{sym}{Symbol of separation (Default ", ")}

\item{quote}{Vector elements to be quoted or not (Default = T)}
}
\value{
Una stringa di elementi formattati al meglio
}
\description{
Funzione che permette partendo da un vettore, di riscrivere quel vettore in varie forme
}
